+++ src/oihana/ftp/FtpClient.php
@@ -3,6 +3,7 @@
 namespace oihana\ftp ;
 
 use oihana\ftp\traits\FtpConnectionTrait ;
+use oihana\ftp\traits\FtpCryptoTrait ;
 use oihana\ftp\traits\FtpDirectoryTrait ;
 use oihana\ftp\traits\FtpFileTrait ;
 
@@ -39,6 +40,7 @@ use oihana\ftp\traits\FtpFileTrait ;
 class FtpClient
 {
     use FtpConnectionTrait ,
+        FtpCryptoTrait ,
         FtpDirectoryTrait ,
         FtpFileTrait ;
 }