+++ tests/oihana/ftp/support/HasFtpClientFixture.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace tests\oihana\ftp\support ;
+
+use oihana\ftp\traits\HasFtpClientTrait ;
+
+/**
+ * A bare host class exercising {@see HasFtpClientTrait}.
+ *
+ * @package tests\oihana\ftp\support
+ */
+class HasFtpClientFixture
+{
+    use HasFtpClientTrait ;
+}